+++ tokenauth/token.go
@@ -47,6 +47,7 @@ type Claims struct {
 	Authenticated bool   `json:"authenticated"`                 // Did the user authenticate? (false on refresh)
 	Service       bool   `json:"service"`                       // Is this token for a service account?
 	Admin         bool   `json:"admin"`                         // Is this token for an admin?
+	System        bool   `json:"system"`                        // Is this token for a system admin?
 
 	// User Data: DO NOT USE AS IDENTIFIER OR SHARE WITH THIRD-PARTY SERVICES
 	Name        string            `json:"name,omitempty"`         // User full name