+++ tokenauth/token.go
@@ -46,6 +46,7 @@ type Claims struct {
 	Anonymous     bool   `json:"anonymous"`                     // Is the user anonymous?
 	Authenticated bool   `json:"authenticated"`                 // Did the user authenticate? (false on refresh)
 	Service       bool   `json:"service"`                       // Is this token for a service account?
+	FirstParty    bool   `json:"first_party"`                   // Is this token used by a first party service (eg. ROKWIRE building block)?
 	Admin         bool   `json:"admin"`                         // Is this token for an admin?
 
 	// User Data: DO NOT USE AS IDENTIFIER OR SHARE WITH THIRD-PARTY SERVICES