+++ sigauth/signature.go
@@ -15,7 +15,6 @@
 package sigauth
 
 import (
-	"bytes"
 	"crypto"
 	"crypto/rand"
 	"crypto/rsa"